Added system for loading pregenerated worlds from JSON
The new system loads JSON generated by another project, which itself is
prototyped as a set of Python tools; these will be integrated and/or
rewritten at some future date

Changes to be committed:
	modified:   ../Cargo.toml
	modified:   camera.rs
	modified:   engine/mod.rs
	modified:   main.rs
	modified:   map.rs
	new file:   mason/json_map.rs
	modified:   mason/mod.rs
	renamed:    engine/image_loader.rs -> mason/rexpaint_loader.rs
	modified:   mason/rexpaint_map.rs
	modified:   sys.rs

+++ src/camera.rs
@@ -206,8 +206,8 @@ pub fn camera_update_system(mut camera:   ResMut<CameraView>,
 				false
 			};
 			// Update the terrain map
-			if map_x >= 0 && map_x < world_map.width
-			&& map_y >= 0 && map_y < world_map.height
+			if map_x >= 0 && map_x < world_map.width as i32
+			&& map_y >= 0 && map_y < world_map.height as i32
 			&& (is_visible || has_seen)
 			{
 				camera.terrain[scr_index] = world_map.tiles[map_index].clone().into();
